implementing useWindowSize() as a custom hook
