+++ src/shared/constants.js
@@ -2,7 +2,7 @@
 /* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
 /* eslint-disable max-len */
 /* eslint-disable no-nested-ternary */
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import {
   Button, Grid, Typography, Tooltip, Box,
 } from '@mui/material';
@@ -1271,3 +1271,34 @@ export const getExpertsData = (councilId) => {
       ];
   }
 };
+
+const useWindowSize = () => { // Initialize state with undefined width/height so server and client renders match
+  // Learn more here: https://joshwcomeau.com/react/the-perils-of-rehydration/
+  const [windowSize, setWindowSize] = useState({
+    width: undefined,
+    height: undefined,
+  });
+
+  useEffect(() => {
+    // Handler to call on window resize
+    function handleResize() {
+      // Set window width/height to state while taking out the width of the horizontal scrollbar
+      setWindowSize({
+        width: window.innerWidth - (window.innerWidth - document.documentElement.clientWidth),
+        height: document.documentElement.clientHeight,
+      });
+    }
+
+    // Add event listener
+    window.addEventListener('resize', handleResize);
+
+    // Call handler right away so state gets updated with initial window size
+    handleResize();
+
+    // Remove event listener on cleanup
+    return () => window.removeEventListener('resize', handleResize);
+  }, []);
+
+  return windowSize;
+};
+export default useWindowSize;
